Don't send NICK messages more than once when we're in channels
The server was updating its state, then checking for channel members
who needed to be informed of the nick change. This was causing users to
get more than one NICK message back whenever they change nicks.

I moved the state-updating part further down, to prevent this. I also
added a test for this.